Remove skija package completely and merge original skiko PaintTest and one ported from skija.
