+++ skiko/src/jvmTest/kotlin/org/jetbrains/skiko/DataTest.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.skiko.skija
+package org.jetbrains.skiko
 
 import org.jetbrains.skia.Data
 import org.junit.Assert.assertArrayEquals
