+++ skiko/src/jvmTest/kotlin/org/jetbrains/skiko/PaintTest.kt
@@ -6,6 +6,8 @@ import org.jetbrains.skiko.util.loadResourceImage
 import org.junit.Assume.assumeTrue
 import org.junit.Rule
 import org.junit.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertNotEquals
 
 class PaintTest {
     @get:Rule
@@ -45,4 +47,40 @@ class PaintTest {
 
         screenshots.assert(surface.makeImageSnapshot())
     }
+
+    @Test
+    fun paintTest() {
+        // TODO: ported from skija and we need address the commented out assertions
+        val paintA = Paint().apply { color = 0x12345678 }
+        val paintB = Paint().apply { color = 0x12345678 }
+
+        assertEquals(paintA, paintB)
+//        assertEquals(paintA.hashCode(), paintB.hashCode(), "hash codes are not identical")
+//
+//        val paintC = Paint().apply { color = -0xcba988 }
+//        assertEquals(paintA, paintC)
+//        assertEquals(paintA.hashCode(), paintC.hashCode())
+//
+
+        assertNotEquals(Paint(), Paint().apply { isAntiAlias = false })
+        assertNotEquals(Paint(), Paint().apply { isDither = true })
+
+        Paint().use { paint ->
+            paint.color = 0x12345678
+//            assertEquals(false, paint == paint.makeClone(), "cloned paint is not equal")
+//            assertEquals(paint, paint.makeClone())
+//            assertNotEquals(paint.hashCode(), paint.makeClone().hashCode())
+        }
+
+        Paint().use { paint ->
+            assertEquals(false, paint.hasNothingToDraw())
+            paint.blendMode = BlendMode.DST
+            assertEquals(true, paint.hasNothingToDraw())
+            paint.blendMode = BlendMode.SRC_OVER
+            assertEquals(false, paint.hasNothingToDraw())
+            paint.alpha = 0
+            assertEquals(true, paint.hasNothingToDraw())
+        }
+    }
+
 }
