+++ skiko/src/jvmTest/kotlin/org/jetbrains/skiko/PathMeasureTest.kt
@@ -1,9 +1,10 @@
-package org.jetbrains.skiko.skija
+package org.jetbrains.skiko
 
 import org.jetbrains.skia.Matrix33
 import org.jetbrains.skia.Path
 import org.jetbrains.skia.PathMeasure
 import org.jetbrains.skia.Point
+import org.jetbrains.skiko.skija.assertCloseEnough
 import kotlin.test.Test
 import kotlin.test.assertEquals
