+++ skiko/src/jvmTest/kotlin/org/jetbrains/skiko/TypefaceTest.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.skiko.skija
+package org.jetbrains.skiko
 
 import org.jetbrains.skia.Data
 import org.jetbrains.skia.FontFamilyName
@@ -6,6 +6,7 @@ import org.jetbrains.skia.FontStyle
 import org.jetbrains.skia.FontVariation
 import org.jetbrains.skia.FontVariationAxis
 import org.jetbrains.skia.Typeface
+import org.jetbrains.skiko.skija.makeFromResource
 import kotlin.test.Test
 import kotlin.test.assertContentEquals
 import kotlin.test.assertEquals
